This commit fixes two bugs: [ 1467537 ] NullPointerException After Close All Files/Open Folder [ 1466590 ] Cancel in "Save?" Dialogs on Quit The first bug was caused by File.listFiles returning null in case of an IO error. I looked for every use of File.listFiles and inserted a check for null. The second bug was mainly handled by doing a round of canAbandonFile calls that take care of saving and possibly cancelling. Only if all files were processed that way, all files were closed.
The commit also removes the uses of getEnclosingClassName, which
unfortunately is incorrect and slow.

M    src/edu/rice/cs/drjava/model/DummyGlobalModelListener.java
M    src/edu/rice/cs/drjava/model/debug/DocumentDebugAction.java
M    src/edu/rice/cs/drjava/model/GlobalModelTestCase.java
M    src/edu/rice/cs/drjava/model/DefaultGlobalModel.java
M    src/edu/rice/cs/drjava/model/definitions/DefinitionsDocumentTest.java
M    src/edu/rice/cs/drjava/model/definitions/DefinitionsDocument.java
M    src/edu/rice/cs/drjava/model/DefaultJavadocModel.java
M    src/edu/rice/cs/drjava/model/GlobalEventNotifier.java
M    src/edu/rice/cs/drjava/model/junit/DefaultJUnitModel.java
M    src/edu/rice/cs/drjava/model/GlobalModelListener.java
M    src/edu/rice/cs/drjava/model/GlobalModelJUnitTest.java
M    src/edu/rice/cs/drjava/model/DummyOpenDefDoc.java
M    src/edu/rice/cs/drjava/model/DummyGlobalModel.java
M    src/edu/rice/cs/drjava/model/AbstractGlobalModel.java
M    src/edu/rice/cs/drjava/model/SingleDisplayModel.java
M    src/edu/rice/cs/drjava/model/OpenDefinitionsDocument.java
M    src/edu/rice/cs/drjava/config/OptionConstants.java
M    src/edu/rice/cs/drjava/ui/MainFrame.java
M    src/edu/rice/cs/drjava/ui/JarOptionsDialog.java
M    src/edu/rice/cs/drjava/ui/config/ConfigFrame.java
M    src/edu/rice/cs/util/FileOps.java
M    src/edu/rice/cs/util/jar/JarBuilder.java


git-svn-id: file:///tmp/test-svn/trunk@3760 fe72c1cf-3628-48e9-8b72-1c46755d3cff

+++ drjava/src/edu/rice/cs/util/FileOps.java
@@ -161,9 +161,12 @@ public abstract class FileOps {
    */
   private static void getFilesInDir(File d, List<File> acc, boolean recur, FileFilter filter) {
     if (d.isDirectory()) {
-      for (File f: d.listFiles(filter)) {
-        if (f.isDirectory() && recur) getFilesInDir(f, acc, recur, filter);
-        else if (f.isFile()) acc.add(f);
+      File[] files = d.listFiles(filter);
+      if (files!=null) { // listFiles may return null if there's an IO error
+        for (File f: files) {
+          if (f.isDirectory() && recur) getFilesInDir(f, acc, recur, filter);
+          else if (f.isFile()) acc.add(f);
+        }
       }
     }      
     else acc.add(d);
@@ -340,7 +343,9 @@ public abstract class FileOps {
 
     boolean ret = true;
     File[] childFiles = dir.listFiles();
-    for (File f: childFiles) { ret = ret && deleteDirectory(f); }
+    if (childFiles!=null) { // listFiles may return null if there's an IO error
+      for (File f: childFiles) { ret = ret && deleteDirectory(f); }
+    }
     
     // Now we should have an empty directory
     ret = ret && dir.delete();
@@ -361,7 +366,9 @@ public abstract class FileOps {
     //  on the directory itself.
     if (dir.isDirectory()) {
       File[] childFiles = dir.listFiles();
-      for (File f: childFiles) { deleteDirectoryOnExit(f); }
+      if (childFiles!=null) { // listFiles may return null if there's an IO error
+        for (File f: childFiles) { deleteDirectoryOnExit(f); }
+      }
     }
   }
 
@@ -405,19 +412,23 @@ public abstract class FileOps {
     while (! working.empty()) {
       PrefixAndFile current = working.pop();
       File [] subDirectories = current.root.listFiles(directoryFilter);
-      for (File dir: subDirectories) {
-        PrefixAndFile paf;
+      if (subDirectories!=null) { // listFiles may return null if there's an IO error
+        for (File dir: subDirectories) {
+          PrefixAndFile paf;
 //         System.out.println("exploring " + dir);
-        if (current.prefix.equals("")) paf = new PrefixAndFile(dir.getName(), dir);
-        else  paf = new PrefixAndFile(current.prefix + "." + dir.getName(), dir);
-        working.push(paf);
+          if (current.prefix.equals("")) paf = new PrefixAndFile(dir.getName(), dir);
+          else  paf = new PrefixAndFile(current.prefix + "." + dir.getName(), dir);
+          working.push(paf);
+        }
       }
       File [] javaFiles = current.root.listFiles(JAVA_FILE_FILTER);
 
-      //Only add package names if they have java files and are not the root package
-      if (javaFiles.length != 0 && !current.prefix.equals("")) {
-        output.add(current.prefix);
+      if (javaFiles!=null) { // listFiles may return null if there's an IO error
+        //Only add package names if they have java files and are not the root package
+        if (javaFiles.length != 0 && !current.prefix.equals("")) {
+          output.add(current.prefix);
 //         System.out.println("adding " + current.prefix);
+        }
       }
     }
     return output;
